Added tests and parsing
So I got as far as trying to write the validation function before I
realized that to really trust RTT as a metric we needed to break out from
hop to hop only advertisement and pass it along as it's own seperate metric.

So off to make that modification to babel now.