guac-light-client: temporarily fix build issues
As of now guac doesn't do all that much, but it still made the firmware
build system a little upset. This commit just moves that out of the way
at a minimum cost so that we're still able to make sure that all
dependencies are build without major hiccups in the firmware.