Optimizations to Reduce binary size
Three major changes have been performed here.

We enable link time optimization, so that all
the libraries we compile are stipped down to
exactly the functions we use.

We enable the 's' optimization flag, which
compiles with size optimization instead of
performance ones.

And we move from the Jemalloc memory allocator
to dynamically linking against the system allocator.

Total reduction in binary size is a little more
than half.