.gitignore: Stop ignoring Cargo.lock
This commit aims to help maintain determinism of althea_rs builds.

The `Cargo.lock` file itself is usually automatically updated in tandem
with dependency changes in a project. The dependencies listed inside are
always in precise versions which fall into the requirements of the
project's crates. These precise versions are supposed to make Cargo
always download the same version of a crate even if its dependency
version in the `[dependencies]` section of a `Cargo.toml` is set to
`"*"`.