WIP config bootstrapping
This contains the new binary Clu which needs to be moved over to a
library as we do the monorepo refacotr. This also adds post_office
very very WIP right now it's a simple request server for the exit
that takes new exit client requests and adds them to the exit database.