Add crosscompile tests
This adds a crosscompilation test for Travis, it pulls a version
of the openwrt toolchain pre built from the Althea updates server
then uses this toolchain to build a proper mips binary. We had
to exclude beta and stable because we use binary shrinking opt
modes which are nightly only.