Force filesystem consistency after edits
This isn't really required for just the files we edit, but on UBIFS
targets like the EdgerouterX we run into issues with other UCI files.
This is perhaps overkill, but forcing a full fs sync should help out.
The alternative is to just put it on cron or actually find the problem
between UBIFS, UCI, and the device NAND itself.