Correct default exit ip range
So the 172 range is 16 not 168 for the private section. We're lucky
this doesn't seem to break many sites because the block is assigned
to AOL and presumably used for Dialup users.

This also expands the assignment range to the maximum allowed, which
is more than a million people on a single exit server, not somthing
I ever expect to see.