Parsing of alpha3 style hellos removed too early
So alpha 3 has what we wanted to change, alpha 4 has both new and old
support but by default it sends old packets. Alpha 5 needs to move
the default to the new packets but keep support for hearing the old
stuff. Then in alpha 6 the change is finally totally done. In the future
we should be more careful to mark these correctly. And/or insitute a
transition date so that the alpha 4 releases moved to future on their own