Explicit Peer Discovery
This adds the PeerListener module which handles peer discovery internally
in pure rust rather than using ip neigh. Also included is a refactor of TunnelManager
to properly take advantage of the optimizations made possible by this.

What's not in this release are similar changes to TrafficWather and optimizations to
Kernel Interface