Added a `rustfmt.toml` file.
Contents of this file works only in `nightly` rustfmt. Hopefully
it will make always a consistent output even if `nightly` rustfmt
becomes stable. Once that will happen we can safely remove `+nightly`
from rustfmt calls with a hope that the settings are consistent.

According to docs `nightly` and `stable` output is different because of
changes in how the syntax is handled.