Correct Stan's Nits
This biggest change here is the move to HashMap's for a few of our
state stores, this provides O(1) lookups which while probably slower
in the common case of 1-2 peers is much faster in the worst case
than iterating over every peer for every other peer during discovery.

Other changes involve no longer cashing when we run out of ports
and instead continuing to operate with a warning.