Added "development" feature switch.
This is off by default, and contains a big warning at the start
of main that this build is for development purposes only and
shouldn't be run on production.

This is an improvement over "debug" build since its not easy to put a
release (lto, etc.) build with debug mode on a router without modifying
Cargo.toml.