Resolve tunnel state disagreement
Before the Peer discovery refactor we would always mutate the existing
tunnel during a successful identity exchange. Now we maintain a tunnel
state and only make modifications when required.

This works fine until one side reboots and the other does not, now there
is disagreement over what the tunnel state is. We need to introduce a single
bit of extra state. We can't do this by not sending hellos when tunnels are
already open. Because in NAT punching cases this prevents tunnel opening totally.

Accepting that we need constant hello sending as a liveness measure we need a way
for one side to indicate to the other that a problem exits. So we send a bool have_tunnel
with our identity, indicating that we do or do not have a tunnel with the peer already.

To complicate matters we don't actually know when we send our hello if we already have
a tunnel, because we send hello's to IP's but open tunnels with identities. To resolve
this bootstrapping problem the bool is wrapped in an option.

So the total system has two bools wrapped in options, that's effectively 16 possible
states. The possible states are reduced to 6 due to these constraints.

Hello responses always include Some(bool) since the hello has an identity we can use for the
lookup.

When a None is recieved we interpret it as 'True' so that we don't delete a tunnel that
might be perfectly fine

Therefore there is only on state in which we act on this bool, which is when we have a
tunnel by the hello response said they do not. In that case we delete our current tunnel
and create a new one matching the local identity sent in the message that's currently being
processed.

With this system no matter what state any side is in one side is garunteed to either be in
the right state already or take action to correct the tunnel state.