rita.sh: Change default revisions to ""
This fixes a design flaw that the integration tests had up until now.
Namely, in CI we would test againsta master and not the PR's feature
branch.

The solution works because git's default behavior is to check out a repo
clone where the remote repo was checked out at.

rita.sh:
* Default REVISION_A to "" and REVISION_B to REVISION_A
* Make "git show" not use a pager and only be issued under VERBOSE

.travis.yml:
* Account for new revision defaults