Correct race condition in wifi meshing
This isn't all this feature needs to work well, but at least we actually
get around to listening on the wlan interfaces. wifi restart actaully takes
nearly 30 seconds to restart the interfaces and fully prep them for networking.
We have to wait until that is done before we even try anything