Refactor interface counter parsers.
Thanks to unit tests the logic there was greatly simplified to use only
a single regexp which is now created once for duration of the program to
amortize the performance of compiling it.

For extra safety the regular expression was improved with named groups
to make sure the correct groups are extracted properly.

It also uses structs instead of a tuple for greater readability in
caller sites.