Update reqwest to converge Tokio
So I was chasing our futures timeout issues and found that Tokio-core
loop poll was taking a dramatic amount of time. But this is an old
tokio library being used by Reqwest, I think having the newer and
the older trying to run in the same event loop was producing bad
results.