Fix link to internet on neighbors page
The entire neighbors screen is filtered through the lense of
your selected exit. In the cases where you don't have a selected
exit or a different error occurs we previously inserted a 0
as the response value. But this is incorrect as 0 is actually the
ideal connection quality. We should instead return u16 max.