Determine Gateway status by parsing iface status
This works well in situations where there is little traffic over the WAN
port. Previously traffic watcher would check the interface counter but now
that things are generally running faster it would often get zero and disable
gateway functions prematurely even if it sometimes got a packet or two.

This method is a more reliable detector that the port is in fact active