Speed up Exits endpoint
This started as a fully async implementation of this endpoint, that quickly
evolved into a huge mess. This much more tame change resolves the actual problem.

Since we ping syncronously with a 1 second timeout the worse case for this endpoint
is when we want to watch it most, during bootstrapping. Sometimes it would dramatically
slowdown startup times when users would monitor the exits page.

This solution is much simpiler, we simply don't run pings when they shouldn't work. This
totally resolves the bootstrapping worst case and prevents useless slowdowns. It doesn't
sovle the fact that this endpoint will eventually become infeasibly slow as we expand
the exits list.