More defensive coding for TrafficWatcher
Today we had a crash in prod because of the TrafficWatcher debts
unwrap. As far as I can tell it's possible for this to happen
if we have a tunnel open for a peer but Babel has yet to insert
any kind of route for them.

This race condition required thousands of hours to reveal, hopefully
with more defensive handling and better logging we can identify rare
but possible conditions faster.