Add remote logging functionality
This adds syslog for remote syslog logging. The primary issue here is that
syslog is unencrypted, so we can't just forward it to a random internet
address, instead we send it to the exit and then configure exits to forward
this data off to the logging aggregation server. This has the nice secondary
effect of making logging more efficient on the aggregation server and securing
the logs since they transit over wg_exit only and they are tls encrypted from
the exit to the aggregator.

Also added are dashboard endpoints
/remote_logging/enabled/{bool}
/remote_logging/level/{0-3}

Sadly a reboot is required to enable log aggregation after using these endpoints
since logging backends can only be initialized once in Rust. Likewise once you
enable remote logging logs no longer appear locally at any point.

These catch-22's are the best solution to the very hard problem of sheer log volume
we can't have an external on router program process the logs because at the trace
level we will overhwelm internal storage in about a minute and memory in less than 10.

By using rust syslog we can send logs directly over udp at the cost of having to restart
the router and not being able to see the logs locally.