Use loglevels in the settings
We have to store it as a freeform string becuase there's no serialize
and deserialize implementation for levelfilter, but we check that
we can parse into level filter on the setting endpoint so we should
be fine.

In the case that an invalid value somehow sneaks in we set to error
logging.