Set logging to INFO when generated temporarily
I'm not super happy about this, but we have 40 somthing routers in
Colombia that are going to come online needing monitoring and because
we use a pull instead of a push arch for updates this is going to control
what their routers do by default.

Current stats harvested from exits let me talk to everyone currently running
Rita and let them know that this will flip on during update so YOLO I guess.