NAT compatible client bandwidth accounting
The previous solution attempted to use iptables to monitor the interfaces
this ended up being a poor idea because it only properly accounted for
traffic sourced from the router. It turns out to be difficult to make
rules that match the NAT rules to count NAT traffic. On the other hand
we can just query the wg_exit interface and keep track of our last seen
bandwidth pretty easily.