Update rita exit traffic watcher to use wireguard counters
The old method seemed to work but was also generally questionable
in my mind since it relied on ipv6 counting and didn't actually
count the traffic we where billing for over the wg_exit tunnel
but instead the peering traffic coming in via ipv6 addresses.

This opened us up to some strange edge cases for example if exits
where to act as an intermediary router at any point it would be
double counted as both intermediary and exit traffic.