Enable Clippy in RLS and try out some fixes
Clippy is a little pedantic and not always correct, which is why I'm not
adding it to CI. This commit just enables clippy as a visible change to RLS
users so that they can look at it's feedback.

Also included are some attempts at using Clippy suggestions, but by no means
a complete set of fixes for all Clippy suggestions in this repo.