Local nonce prediction
When trying to send transactions very quickly we run into the problem
of not getting an updated nonce right away. This comment increments the
nonce after each successful payment and does not allow it to be decreased
by the full node, so that it will remain accurate even if the full node
has not yet seen the transaction.