Validate that transactions are actually real
This adds transaction validation to the payments system, when we
get a payment from a peer we will go and look it up on our own
full node and check that it satisfies all of the properties
required to make it a valid payment to us for the amount expected.

The biggest issue with this commit is that it adds more failure
conditions to a payment being accepted but there's still no resolution
to how to re-converge on lost payments or otherwise 'try again' and
make sure things went through.

I think a good solution might be some sort of watcher that will retry
failed txid validations a few times in the case of some spurious failure
that should reduce things to malicious behavior and uncommon enough edgecases
as to not be an issue.