Standardize on i128 for billing computation
So the recent change to use u64 for exit billing fixes our new speed
problem, but we do in fact need signed integers for these operations.
Which should have been obvious since the Int256 operations I was trying
to optimize replaced i64 operations.

That mistake aside the number of type changes throughout the billing process
is somewhat concerning. Even if all the conversions are safe and the values
are reasonable.

We collect billing data from the system counters as u64, which is reasonable
as long as you running billing for less than 16 exabytes of traffic.

Multiplying by price adds several orders of magnitude thanks to wei, but a i128
provides enough range to charge trillions of eth.

128 integer support is pretty common on Arm and x86 so we should maintain native
speed across most platforms. I think this is the best compromise for now but I
also think future refactors may want to reconsdier if we can make 256 bit integers
fast enough to stop worrying about all of this.