Serde NoDeserialize causes strange cross-thread behavior
NoDeserialize values would be reset to their defaults when you tried
to call SETTING from another thread.

I suspect this is an interaction of various components 'working as intended'
in odd ways. But I haven't managed to root cause it, instead we're just going
to reduce the config saving rate and move to storing these values