Add Oracle and auto pricing updates
This adds the Oracle module which is used for all the blockchain requests
that used to clutter up rita loop. It also adds functionaltiy for a very
simple pricing oracle, it simply hits a centralized url to get pricing updates
in the future these same endpoints and module will contain a real decentralized
auto price adjustment algorithm.