Enforce maximum price
This enforces a maximum price value (optionally updated by oracle) when
the maximum price for a route is reached the max price will be paid.

In the case of chains of intermediaries that have this price they will
end up kicking people to the free tier but it should protect from hostile
pricing.

In the future this value should be moved to the subnet dao