Change cutoff computation
Cutoff is now an even 4 playments rather than the multiplier amount
as you get into smaller fee amounts (5% and 1%) the cutoff threshold
becomes progressivly more ridiculous. It comes down to how long you
think it should take to confirm a transaction and the chances of a
transaction being left out in the cold.

I think this is viable for now, where this sets the effective cutoff
at $2