Restart Rita after mode change
Changing the mode may change the fe80 addresses which will in turn cause
the tunnel the peer had to be dropped eventually, Rita never picks up the
new address and connectivity is lost. This should resolve that even if
it's a bit of a hacky way to go about it.