More robust mesh toggle
This commit removes the possiblity of a config inconsistency when adding
a mesh port, previously due to the mandatory wait when getting new fe80
addresses on a new mesh port there was a 60 second period during which
the router could be shut down and have a mesh port not work properly.

This adds a redundant operation that should prevent that from ever
happening.