Remove client update
I don't understand why update client did what it did, essentially it
let the client update some details about it's registration but none
of it was actually required for things to work and in general it presented
a database security risk. It's now been reduced to routine entry update operations
like bumping the last seen timestamp.