Fix build tests to work with postgres
the cross docker container does not contain postgres libs while the rust_musl_builder
container does, so we move the build tests to cross building only Rita.

This also upgrades our travis environment to allow for native building again, which
is how we test rita_exit and build for the integration tests.

This commit does not fix the integraiton tests, since they depend on netnamespaces
and now we need outside of namespace access for postgres. That will come in another
commit.