Exit server refactor, kill db_client
db_client, being a single actor for all database operations was fundamentally
serializing. This presented a large problem when it came to the setup and status
endpoints, both of which need to hit a db (now potentially many hundreds of ms away).

This would cause the rita exit loop to take more than 5 seconds to complete, since iterations
are spawned in real time but must wait on the previous run the loop would become consistently
slower forever.

This refactor moves the setup and status endpoints to their own worker threads using actix_web
and has the remaining loop members all use a single database connection and clients_list for a
given round.