Add fixed low balance treshold
While having a dynamic payment limit based on gas prices works quite
well for payments it's not a good thing to base a low balance system off
of. What will happen is that the system will reach a state where as
gas prices fluctuate the user goes on and off of the low balance mode
with each block.

A fixed value resolves this problem, but will require future adjustment
and probably oracalization