No more free tier billing discount
The free teir billing discount makes it very difficult to maintain
billing consensus over long periods. Each device will run at it's own
loop speed and a faster or slower sampling rate will result in different
values being discounted under the free tier.

Even if we removed the implied clock operation and added an explicit time
counter a router that sampled every 5 seconds versus every 4 due to whatever
environmental or device load factors will come back with a different billing value.

This commit makes billing easily deterministic, in exchange though the free tier
is now paid out by the client themselves. Considering that we have a pretty large
amount of trust before we initiate throttling in the first place it may simply be
better to cut people off for a while. But that's outside the scope of this commit.