Add fee multiplier to the oracle
During testing today I realized that the fee multiplier held implicit
info about the price, if the transaction fee value versus the fee multiplier
versus the amount of bandwidth being sold is off then it's possible to blow
through the grace period in less time than it may take for a block to be produced
on chain.

Making any one of these factors tunable is pretty equal so I chose the fee multiplier.
This should make it easy for us to keep the price to fee multiplier ratio reasonable.