Make payment in flight state explicit
This refactors debt keeper to make the payment in flight state explicit.

Previously this state was implicit and handled by the payment failed flow
but in light of how many problems that caused during high failure situations
we're going to make it explicit. Hopefully this will both reduce the failure
rate of payments and result in more consistent internal state.