Validate that payments are really on chain
Previously we assumed that once we had a valid txid payments would eventually
reach the blockchain, that's not only not true, it's not true often enough
to cuase a whole host of problems.

Now both the payer and payee watch transactions and wait until they have 4
confirmations before declaring them as paid. Hopefully this will eliminate
disagreements about what payments are actually valid.