Have peer listener wait for interfaces to be active
This one is actually pretty strange, so the routers using their fake
switches typcially always have their mesh interfaces marked as 'up'
with fe80 addresses, but this is an artifect of the switches on
a system with independent interfaces we would at one point simply not
listen until a reboot had occured. Or in the case of more recent versions
crash.

This is a better handling situation in general.