Tweaks to exit-client debts
Some operations in the client loop could be replaced with do_send()
becuase we didn't observe the response and they where protected from
failures already. This simplifies the call flow and produces fewer useless
closures.

The real tweak here is that the exit debts request is free to take quite some time
(up to 10 seconds seemed optimal) to ask the exit for debts. The exit takes far
longer than it should to reply to these requests due to the syncronous nature of
it's main loop and database operations. It still succeeds often enough to be fine
for now. With this new structure client traffic watcher can safely wait indefinately
without causing issues but updating with a very old debt value is risky, so we timeout
at 5 seconds.