Revert TrustDNS update
This is mostly about reverting the associated deps update, which
seemed to cause a spooky action at a distance sort of bug with
the exit network manager. The program would exit with 255 with no
stack trace or other visible error when a user requested status or
setup from the exit. It still seems that we didn't need the actix_utils
module so that's staying out