Add back client TrafficWatcher for bandwidth usage count
We can't really determine exactly how much we have used by asking
the exit for a bill. I'm also hesitant to have the exit transfer
even more data about the user. The more data is provided the larger
the vulnerability surface there.

So the easiest way to solve this problem is to revive Client traffic
watcher and update the comments to reflect how it is now used.