Lower payment timeout again
We had an incident last night where the gateway was waiting for a timed
out tx and failed to pay a relay in a timely manner, resulting in service
disruption.

Trying to reduce the load on the blockchain does not seem to be functionality
frendly neither does it seem to be what full nodes encourage, I've never seen
a tx get in after more than a couple of minutes.