Rename connection pool setting to workers
It's difficult to take a connection pool size and assign one less
endpoint worker. Because it seems we need a minimum of two to run quickly
enough to even pass our own tests. This commit changes the var name so
that it makes more sense. Rather than connection_pool_size being the
number of workers and one less than the actual pool size