Resolve race condition where the exit may be paid twice
Since we ask the exit what it believes we owe there is the possibility
of a race condition, where we have a payment in flight but the exit has
not updated it's debts. This commit resolves this problem by recording
when we last made a successful payment and having the local nodes payment
updates wait to update the debt to the exit while a payment is in flight
or if a payment has just been completed.