Remove bounty hunter and rebuild lockfile
The bounty hunter was constraining the build for diesel and other dependencies
to try and find matching versions. This isn't great becuase we mostly are not using
Bounty hunter right now and as a totally different binary keeping the same deps versions
is actually not at all important. So I'm breaking it out into another repo.