Properly handle buffer exhaustion
First off there was a very obvious error in parsing where we
never tried to parse and return the fully assembled message.

That was the core of the issue here, the only other change is to
detect buffer exhaustion, recurse right away and eliminate it from the
timeout. This should let us handle arbitrarily large babel dumps up
until we actually run out of system memory.