Move Oracle into DAO settings
The concept of the price oracle as a currency based function doesn't
really reflect how it's used. It's mostly used as a placehold for the
DAO's ability to roll out agreed upon settings changes to a network.

This move reflects that more directly and will solve obvious problems
like some networks wanting dramatically different prices in the same
currency.