Implement toggleable debt limit
Enforcing a large unpaid debt is somewhat difficult in a trustless system
for obvious reasons. While we've made great strides with getting devices to
remember and agree on what these debts are that doesn't make the money required
to pay them come out of nowhere.

Before this patch if the user was on the 'free tier' (1mbps limited connection allowed
when the user has not paid) for a long time and deposited money a lot of it may go poof
to get the system back online at full speed by paying off all debts.

This is an attempt to limit that to a reasonable amount to prevent balances going poof
when a router was run dry for long periods.