Exempt the /exits endpoint from authentication
/exits is a pretty safe read only endpoint, that does reveal some
info about the debugging state of the router but not anything else
appreciable. Specifically we query this endpoint do determine if the
router is in a good condition in the checkup script.

Adding the password to that script when we set one is nearly infeasible,
so the only other answer would be to create a seperate server (using more
ram and another thread handle) to provide a no-auth endpoint to check against.

For various reasons that would be somewhat risky to roll out and this seems to
be the better strategy, even if it is a little hacky