More aggressive tunnel timeout
In the case where a router is restarting consantly on the cron timeout
the 10 minute cron job will eventually exhaust the tunnel supply of upstream
routers. This settings tweak will allow us to maintain functionality in that
situation indefinately. Now if that's acutally a good idea remains to be seen.