Use the 'metro' tuning for Cake by default
Most of our links have a average latency between 1ms and 10ms, the default
'internet' value of the per hop tunnel Codel latency target (100ms) was
grossly off base.

This was hidden by the shaping on exit tunnels (which worked) and the shaping
for throughput constrained devices (which also worked). This issue was only
encountered when we had a throughput limited link that was saturated by more
than a single exit tunnel at a time.