Handle balances returning to zero
This is a pretty obscure special case. Because full nodes may return
zero as a balance if they are still syncing and it very much disturbs
users to have their balance go to zero for even a few seconds we ignore
zero inputs for balances once our balance has in the past been nonzero.

Of course if you use the withdraw_all endpoint and intentionally zero your
balance this presents a problem. The router will from there on out never show
your accurate zero balance. This adds a simple message sent to the oracle
from the withdraw_all endpoint that allows for the router balance to return
to zero for a short window.