Add qdisc speed limitations to the Neighbors endpoint
These should be displayed here for the sake of allowing users to
see what's going on with their interfaces.

This patch also changes the semantics of the neighbors endpoint to
include only people you have active tunnels with, not everyone you
have debts for. Since we persist debts longer term now that was having
unintended side effects