Skip deserializing payment_in_flight
If a router had happened to save it's debts while a payment was in flight
and was then rebooted before that data could be overwritten the payment_in_flight
bool would forever be stuck to true prevent the payment of that spefific node.

By not deserializing this variable we can ensure it's always set to false on reboot
this may in some rare cases result in a double payment, but that's better than no
payment.