Go back to forgiving debts on reboot
This change has me somewhat conflicted, obviously if you want a trustless billing
system nodes should remember what other nodes owe them and hold them to it.

On the other hand when bugs occur with persistant data within routers it's very very
difficult for people IRL to debug. After a few months without this feature we finally
ran into an issue post Xdai transition which is understandable since the monetary
unit has changed behind the scenes. But no matter how understandable it is techncially
it's not acceptible to users who accept the loss of a few dollars much more redily
than software debugging quests.

So for now I'm putting this back in. It's definately the sort of thing that you would
rather not have 99.99% of the time, but that 0.001% is capable of souring a users Althea
experience totally.