Ipv6 netmask considerations
With ipv4 we are only assigning a single ip per user which means
we can have one global netmask and everything works out. But since we
are assigning subnets now we need to consider building a proper routing
table for these subnets, meaning we can't simply apply a /32 to the wg_exit
interface and call it a day.

Instead we need to specify a route to a subnet via the starting address
which the router assigns to itself