Deal with unassigned client ipv6
In most cases we are handling registration and we can assume that ipv6 was properly
generated.The only real tricky case is old client checkin where we have to add a check
to generate them an ipv6 address to deliver via an update.

From the clients perspective post exit update they will get a Some() value for their
exit checkin and integrate it into their own config.