Don't read opkg feeds file if there is no update
It turns out that match isn't evaluated like an if a() && b() where b
would not be run if a evaluates to false. So I've changed the call
structure to avoid get_release_feed() from being called when there
is no update to compare it to. Otherwise we'll be doing blocking IO
in our event loop every 5 seconds.