Make ipv6 for exits totally optional
After looking at the ipv6 support of various hosting options I've
concluded that it will often be difficult to host an ipv6 compatible
exit and that upgrading our existing exit fleet to be ipv6 compatible
across the board will be difficult and time consuming.

So we at least need to be able to handle non ipv6 exits for the forseeable
future.