Fix crash when client can not use free tier
The transform we use not to re-add already existing rules failed on a
delete command. This is easily fixed as we don't need the same check flow
for our delete commands right now.

I'd like to consider phasing that out entierly and only managing current iptables
state from Rita, but there are issues with that like what to do when Rita gets
restarted and has to deal with parital iptables state