Assign unique link local ipv6 address per interface
The issue is that tunnels to the same destination over multiple interfaces
would drop packets and generally fail to work outside of the ifidx zero case.

To the best of my understanding this is caused by multiple levels of ambiguity
within the kernel over IpV6 link local addresses (fe80), WireGuard handshakes
and the way the kernel processes packets.

We set the same fe80 address (generated using the last 64 bits of our mesh ip)
on every physical interface. This means that packets delivered for address X
might end up delivered to different Wireguard tunnels on different interfaces
since the specific scope details of fe80 addresses aren't accounted for in many
places this would be accepted WireGuard would consdier it a handshake and you
would now have a tunnel listening on physical interface X handshaked with a tunnel
that was trying to send packets to phyiscal interface y.

This leads to packet loss and other poor behavior as these handshakes are redone
over and over.

The solution in this patch is to xor the ifidx with the last two bytes of the
link local ipv6 address, providing a unique address to every interface.

This does introduce one new potential edge case, Babel will now see two
neighbors on one interface becuase the router on the other end is now
responding from two addresses on that wg interface. I'm not sure this is
resolveable.