Add light client tunnels
This commit creates a second tunnel type in parallel to the existing
per hop tunnels. This new 'light client tunnel' is a variant of the
existing tunnels that is attached to a bridge interface internally
and serves up a ipv4 address to NAT over this bridge.

This new tunnel type is only accessible by sending a hello to a different
port, this port is intended to be firewalled from the larger mesh network.

What this setup ultimately does is facilitate the use of dumb ipv4 devices
as clients. It will also require a billing implementation very similar to
the one currently used for Rita exit