Fix default port numbering
So it looked like 4874 was open but in fact it's being used as
the production hello port. The defualt section of the config was
not synced with the actual templated config we deploy in the firmware.

This updates that and fixes the light client hello port to the next
actually unused port. We could use the exit registration port 4875
but I decided against that.